Used CrawlerDetect library to prevent popup survey from being displayed to robots and crawlers.  It's unlikely to be 100% effective but at a minimum it should handle the big players.  I modified the library to remove the php namespace statements to comply with the Drupal way.
The CrawlerDetect library will likely need to be updated periodically to
get new user agent signatures.

ALso fixed a bug where an error message was displayed to the user when
they attempted to configure the popup survey before an appropriate block
was created.  Config page now displays 'None' as the only option.