Updated config screen to make the bot list editable and added basic input validation on blacklist value.  Blacklis is now passed to the JS using the Drupal_add_js mechanism.
Fix for issue 1383.  Bot/crawler detection occurs on client
side using most common known bots and crawlers.  This will not exclude
unknown bots but could be used exclude specific problematic browsers or
mobile devices.

Made minor change to have the survey panel be hidden by default and only
shown when needed.  This fixed a screen flicker that happened when the
popup was rendered and then hidden.